pyproject.toml
[project]
name = "job-analyzer-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow[amazon,postgres]==2.10.0",
    "boto3>=1.42.27",
    "dotenv>=0.9.9",
    "duckdb>=1.4.3",
    "openai>=2.15.0",
    "requests>=2.32.5",
    "spacy>=3.8.11",
]
